api/requirements.txt
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-openapi3==4.1.0
flask-openapi3-swagger==5.21.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
nodeenv==1.9.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyright==1.1.403
python-dotenv==1.1.1
ruff==0.11.13
SQLAlchemy==2.0.41
sqlmodel==0.0.24
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
